requirements.txt
pandas
numpy
firebase-admin
google-cloud-firestore
yt_dlp
python-dotenv
beautifulsoup4
requests
feedparser
python-dateutil # Added for robust date parsing
flask # For Cloud Run web service
gunicorn # For production WSGI server
google-genai # For LLM processing
langchain
langgraph
httpx # For downloading files from URLs
langchain-google-genai